+++ src/options/components/App/App.tsx
@@ -30,23 +30,26 @@ const getContent = (
     authenticated: boolean,
     requestProcessState: RequestStatus,
     isPremiumToken: boolean,
+    isContentLocked: boolean,
 ) => {
     if (authenticated) {
         return (
             <div className="wrapper">
                 <Sidebar />
-                <div className="content">
-                    <Switch>
-                        <Route path="/" exact component={Settings} />
-                        <Route path="/exclusions" exact component={Exclusions} />
-                        <Route path="/account" component={Account} />
-                        <Route path="/about" component={About} />
-                        <Route path="/support" component={Support} />
-                        {!isPremiumToken && (
-                            <Route path="/free-gbs" component={FreeGbs} />
-                        )}
-                        <Route component={Settings} />
-                    </Switch>
+                <div className="content" inert={isContentLocked ? '' : undefined}>
+                    <div className="content__wrapper">
+                        <Switch>
+                            <Route path="/" exact component={Settings} />
+                            <Route path="/exclusions" exact component={Exclusions} />
+                            <Route path="/account" component={Account} />
+                            <Route path="/about" component={About} />
+                            <Route path="/support" component={Support} />
+                            {!isPremiumToken && (
+                                <Route path="/free-gbs" component={FreeGbs} />
+                            )}
+                            <Route component={Settings} />
+                        </Switch>
+                    </div>
                 </div>
             </div>
         );
@@ -65,6 +68,7 @@ export const App = observer(() => {
         authStore,
         settingsStore,
         globalStore,
+        uiStore,
     } = useContext(rootStore);
 
     useMessageHandler();
@@ -77,11 +81,32 @@ export const App = observer(() => {
 
     const { isPremiumToken } = settingsStore;
 
+    const { isContentLocked } = uiStore;
+
     useEffect(() => {
         (async () => {
             await globalStore.init();
         })();
     }, []);
+
+    /**
+     * We are adding locked class to body in order to lock scrolling in case
+     * if any modal, sidebar is open. We are adding specifically to body because
+     * in Apple devices "spring scroll" is not works if we add locking to div elements.
+     */
+    useEffect(() => {
+        const BODY_LOCK_CLASS = 'locked';
+
+        if (isContentLocked) {
+            document.body.classList.add(BODY_LOCK_CLASS);
+        } else {
+            document.body.classList.remove(BODY_LOCK_CLASS);
+        }
+
+        return () => {
+            document.body.classList.remove(BODY_LOCK_CLASS);
+        };
+    }, [isContentLocked]);
 
     // show nothing while data is loading
     if (status === RequestStatus.Pending) {
@@ -92,7 +117,7 @@ export const App = observer(() => {
 
     return (
         <HashRouter hashType="noslash">
-            {getContent(authenticated, requestProcessState, isPremiumToken)}
+            {getContent(authenticated, requestProcessState, isPremiumToken, isContentLocked)}
             <Notifications />
             <Icons />
         </HashRouter>