+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Ability to add a custom DNS server by pressing the Enter key [#180].
+- UI support for the options page on mobile screens.
 
 ### Fixed